+++ mongospecs/helpers/empty.py
