+++ mongospecs/mixins/signal.py
@@ -1,0 +1,13 @@
+from blinker import signal
+import typing as t
+
+class SignalMixin:
+    @classmethod
+    def listen(cls, event: str, func: t.Callable[..., t.Any]) -> None:
+        """Add a callback for a signal against the class"""
+        signal(event).connect(func, sender=cls)
+
+    @classmethod
+    def stop_listening(cls, event: str, func: t.Callable[..., t.Any]) -> None:
+        """Remove a callback for a signal against the class"""
+        signal(event).disconnect(func, sender=cls)
